requirements.txt
opencv-python
mediapipe
numpy
customtkinter
mysql-connector-python
fpdf
matplotlib
pyttsx3
python-dotenv